created validation environment/script, tested our two models
Created a validation environment that uses new data not included in training, both models kinda shit the bed a bit. I think I need to re-evalauate my reward function, since my models clearly aren't understanding what to do to increase reward. Also, this might be a fluke but my models seem to do well at the beginning of the episode, which could be because they have much more experience at that part of the episode - maybe I should initialize episodes on a random point in my data to train on all my data