tested v10
Decided to scale down my testing and just train my agent to train BTC, to see if there's a problem with the dimensionality or if there are bigger problems afoot. A few things - the strategy the agent developed was to just HODL forever, which isn't optimal.  Maybe this developed as a result of only trading one currency, I'm not sure. Also, it didn't pass validation, which suggests to me that my reward function or something likely needs to change.